Catlike, Object Management, Read/write override in Game class

+++ Assets/General/CatlikeCoding/ObjectManagement/Scripts/GameDataReader.cs
@@ -15,7 +15,7 @@ public class GameDataReader
     {
         return reader.ReadSingle();
     }
-    public float ReadInt()
+    public int ReadInt()
     {
         return reader.ReadInt32();
     }
